Disabled Vulkan on 32-bit windows

+++ Graphics/GraphicsEngineVulkan/include/VulkanUtilities/VulkanMemoryManager.h
@@ -171,7 +171,7 @@ public:
         m_CurrAllocatedSize (rhs.m_CurrAllocatedSize),
         m_PeakAllocatedSize (rhs.m_PeakAllocatedSize)
     {
-        for(int i=0; i < m_CurrUsedSize.size(); ++i)
+        for(size_t i=0; i < m_CurrUsedSize.size(); ++i)
             m_CurrUsedSize[i].store(rhs.m_CurrUsedSize[i].load());
     }
